+++ src/clone_node.ts
@@ -1,8 +1,8 @@
 import { Maybe, maybeOf } from '@fluss/core';
 
-export function cloneNode(
-  node: Node | Maybe<Node>,
+export function cloneNode<T extends Node>(
+  node: T | Maybe<T>,
   deep: boolean = false
-): Maybe<Node> {
-  return maybeOf(node).map((node) => node.cloneNode(deep));
+): Maybe<T> {
+  return maybeOf(node).map((node) => node.cloneNode(deep) as T);
 }