+++ src/has_attribute.ts
@@ -4,6 +4,14 @@ import type { AttributeNamesOf, GlobalAttributeNames } from './types';
 export function hasAttribute<E extends Element>(
   element: E | Maybe<E>,
   name: AttributeNamesOf<E> | GlobalAttributeNames
+): boolean;
+export function hasAttribute<E extends Element>(
+  element: E | Maybe<E>,
+  name: string
+): boolean;
+export function hasAttribute<E extends Element>(
+  element: E | Maybe<E>,
+  name: AttributeNamesOf<E> | GlobalAttributeNames | string
 ): boolean {
   return maybeOf(element)
     .map((el) => el.hasAttribute(name))