+++ src/remove_attribute.ts
@@ -5,6 +5,14 @@ import type { AttributeNamesOf, GlobalAttributeNames } from './types';
 export function removeAttribute<E extends Element>(
   element: E | Maybe<E>,
   name: AttributeNamesOf<E> | GlobalAttributeNames
+): void;
+export function removeAttribute<E extends Element>(
+  element: E | Maybe<E>,
+  name: string
+): void;
+export function removeAttribute<E extends Element>(
+  element: E | Maybe<E>,
+  name: AttributeNamesOf<E> | GlobalAttributeNames | string
 ): void {
   if (hasAttribute(element, name)) {
     maybeOf(element).map((el) => el.removeAttribute(name));